+++ src/fparser/two/Fortran2008/critical_stmt_r811.py
@@ -1,7 +1,7 @@
 # -----------------------------------------------------------------------------
 # BSD 3-Clause License
 #
-# Copyright (c) 2018-2023, Science and Technology Facilities Council.
+# Copyright (c) 2023, Science and Technology Facilities Council.
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -33,48 +33,46 @@
 # -----------------------------------------------------------------------------
 
 """
-    Module containing Fortran2008 Type_Declaration_Stmt rule R501
+    Module containing Fortran2008 Critical_Stmt rule R811
 """
-from fparser.two.Fortran2003 import Type_Declaration_Stmt as Type_Declaration_Stmt_2003
+from fparser.two.utils import StmtBase, WORDClsBase
 
 
-class Type_Declaration_Stmt(Type_Declaration_Stmt_2003):  # R501
+class Critical_Stmt(StmtBase, WORDClsBase):
     """
-    Fortran 2008 rule 501.
+    Fortran 2008 Rule R811.
 
-    .. code-block:: fortran
-
-        type-declaration-stmt is declaration-type-spec [ [ , attr-spec ] ... :: ]
-                                 entity-decl-list
-
-    The implementation of this rule does not add anything to the Fortran 2003
-    variant but overwrites :py:meth:`get_attr_spec_list_cls` to use
-    the Fortran 2008 variant of :py:class:`Attr_Spec_List`.
-
-    Associated constraints are:
-
-    "C501 (R501)  The same attr-spec shall not appear more than once in a given
-          type-declaration-stmt."
-    "C502 (R501)  If a language-binding-spec with a NAME= specifier appears,
-          the entity-decl-list shall consist of a single entity-decl."
-    "C503 (R501)  If a language-binding-spec is specified, the entity-decl-list
-          shall not contain any procedure names."
-    "C505 (R501)  If initialization appears, a double-colon separator shall
-          appear before the entity-decl-list."
-
-    C501-C503, C505 are currently not checked - issue #259.
+    critical-stmt is [ critical-construct-name : ] CRITICAL
 
     """
+
+    subclass_names = []
+    use_names = ["Critical_Construct_Name"]
 
     @staticmethod
-    def get_attr_spec_list_cls():
-        """Return the type used to match the attr-spec-list
+    def match(string):
+        """
+        Attempts to match the supplied string as a CRITICAL statement.
 
-        This overwrites the Fortran 2003 type with the Fortran 2008 variant.
+        :param str string: the string to attempt to match.
+
+        :returns: 2-tuple containing the matched word "CRITICAL" and None or \
+                  None if no match.
+        :rtype: Tuple[str, NoneType] or NoneType
 
         """
-        # Avoid circular dependencies by importing here.
-        # pylint: disable=import-outside-toplevel
-        from fparser.two.Fortran2008 import Attr_Spec_List
+        return WORDClsBase.match("CRITICAL", None, string)
 
-        return Attr_Spec_List
+    def get_start_name(self):
+        """
+        :returns: the name associated with the start of this CRITICAL region (if any)
+        :rtype: str | NoneType
+        """
+        return self.item.name
+
+    def tostr(self):
+        """
+        :returns: the string representation of this node.
+        :rtype: str
+        """
+        return "CRITICAL"
