+++ src/fparser/two/Fortran2008/declaration_construct_c1112.py
@@ -1,7 +1,7 @@
 # -----------------------------------------------------------------------------
 # BSD 3-Clause License
 #
-# Copyright (c) 2018-2023, Science and Technology Facilities Council.
+# Copyright (c) 2023, Science and Technology Facilities Council.
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -33,48 +33,25 @@
 # -----------------------------------------------------------------------------
 
 """
-    Module containing Fortran2008 Type_Declaration_Stmt rule R501
+    Module containing Fortran2008 Declaration_Construct constraint C1112
 """
-from fparser.two.Fortran2003 import Type_Declaration_Stmt as Type_Declaration_Stmt_2003
+from fparser.two.Fortran2003 import Declaration_Construct
 
 
-class Type_Declaration_Stmt(Type_Declaration_Stmt_2003):  # R501
-    """
-    Fortran 2008 rule 501.
+class Declaration_Construct_C1112(Declaration_Construct):  # C1112
+    """Fortran 2008 constraint C1112
+    C1112 A submodule specification-part shall not contain a
+    format-stmt, entry-stmt, or stmt-function-stmt.
 
-    .. code-block:: fortran
-
-        type-declaration-stmt is declaration-type-spec [ [ , attr-spec ] ... :: ]
-                                 entity-decl-list
-
-    The implementation of this rule does not add anything to the Fortran 2003
-    variant but overwrites :py:meth:`get_attr_spec_list_cls` to use
-    the Fortran 2008 variant of :py:class:`Attr_Spec_List`.
-
-    Associated constraints are:
-
-    "C501 (R501)  The same attr-spec shall not appear more than once in a given
-          type-declaration-stmt."
-    "C502 (R501)  If a language-binding-spec with a NAME= specifier appears,
-          the entity-decl-list shall consist of a single entity-decl."
-    "C503 (R501)  If a language-binding-spec is specified, the entity-decl-list
-          shall not contain any procedure names."
-    "C505 (R501)  If initialization appears, a double-colon separator shall
-          appear before the entity-decl-list."
-
-    C501-C503, C505 are currently not checked - issue #259.
+    This class specialises 'Declaration_Construct' to remove
+    'Format_Stmt', 'Entry_Stmt' and 'Stmt_Function_Stmt'
 
     """
 
-    @staticmethod
-    def get_attr_spec_list_cls():
-        """Return the type used to match the attr-spec-list
-
-        This overwrites the Fortran 2003 type with the Fortran 2008 variant.
-
-        """
-        # Avoid circular dependencies by importing here.
-        # pylint: disable=import-outside-toplevel
-        from fparser.two.Fortran2008 import Attr_Spec_List
-
-        return Attr_Spec_List
+    subclass_names = Declaration_Construct.subclass_names[:]
+    subclass_names.remove("Format_Stmt")
+    subclass_names.remove("Entry_Stmt")
+    # Commented out Stmt_Function_Stmt as it can falsely match an
+    # access to an array or function. Reintroducing statement
+    # functions is captured in issue #202.
+    # subclass_names.remove('Stmt_Function_Stmt')
