+++ src/fparser/two/Fortran2008/error_stop_stmt_r856.py
@@ -1,7 +1,7 @@
 # -----------------------------------------------------------------------------
 # BSD 3-Clause License
 #
-# Copyright (c) 2018-2023, Science and Technology Facilities Council.
+# Copyright (c) 2023, Science and Technology Facilities Council.
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -33,48 +33,35 @@
 # -----------------------------------------------------------------------------
 
 """
-    Module containing Fortran2008 Type_Declaration_Stmt rule R501
+    Module containing Fortran2008 Error_Stop_Stmt rule R856
 """
-from fparser.two.Fortran2003 import Type_Declaration_Stmt as Type_Declaration_Stmt_2003
+from fparser.two.Fortran2003 import Stop_Code
+from fparser.two.utils import StmtBase, WORDClsBase
 
 
-class Type_Declaration_Stmt(Type_Declaration_Stmt_2003):  # R501
+class Error_Stop_Stmt(StmtBase, WORDClsBase):  # R856
     """
-    Fortran 2008 rule 501.
-
-    .. code-block:: fortran
-
-        type-declaration-stmt is declaration-type-spec [ [ , attr-spec ] ... :: ]
-                                 entity-decl-list
-
-    The implementation of this rule does not add anything to the Fortran 2003
-    variant but overwrites :py:meth:`get_attr_spec_list_cls` to use
-    the Fortran 2008 variant of :py:class:`Attr_Spec_List`.
-
-    Associated constraints are:
-
-    "C501 (R501)  The same attr-spec shall not appear more than once in a given
-          type-declaration-stmt."
-    "C502 (R501)  If a language-binding-spec with a NAME= specifier appears,
-          the entity-decl-list shall consist of a single entity-decl."
-    "C503 (R501)  If a language-binding-spec is specified, the entity-decl-list
-          shall not contain any procedure names."
-    "C505 (R501)  If initialization appears, a double-colon separator shall
-          appear before the entity-decl-list."
-
-    C501-C503, C505 are currently not checked - issue #259.
+    Fortran 2008 rule R856
+    error-stop-stmt is ERROR STOP [ stop-code ]
 
     """
+
+    subclass_names = []
+    use_names = ["Stop_Code"]
 
     @staticmethod
-    def get_attr_spec_list_cls():
-        """Return the type used to match the attr-spec-list
+    def match(string):
+        """Check whether the input matches the rule
 
-        This overwrites the Fortran 2003 type with the Fortran 2008 variant.
+        :param str string: Text that we are trying to match.
+
+        :returns: None if there is no match or, if there is a match, a \
+            2-tuple containing a string matching 'ERROR STOP' and an \
+            instance of :py:class:`fparser.two.Fortran2003.Stop_Code` \
+            (or None if an instance of 'Stop_Code' is not required and \
+            not provided).
+        :rtype: (str, :py:class:`fparser.two.Fortran2003.Stop_Code` or None) \
+            or NoneType
 
         """
-        # Avoid circular dependencies by importing here.
-        # pylint: disable=import-outside-toplevel
-        from fparser.two.Fortran2008 import Attr_Spec_List
-
-        return Attr_Spec_List
+        return WORDClsBase.match("ERROR STOP", Stop_Code, string)
