+++ src/fparser/two/tests/fortran2008/test_submodule_stmt_r1117.py
@@ -90,31 +90,21 @@ def test_simple_error6():
     assert "Submodule_Stmt: 'submodule (id) (name)'" in str(excinfo.value)
 
 
-def test_simple_error7(monkeypatch):
+def test_simple_error7():
     """Test the parsing of a submodule statement when there is a single
-    right hand bracket. The error generated here is unreachable if
-    splitparen works correctly so we need to monkeypatch.
+    right hand bracket.
 
     """
-
-    monkeypatch.setattr(
-        "fparser.common.splitline.splitparen", lambda x: ["", "id)", "name"]
-    )
     with pytest.raises(NoMatchError) as excinfo:
         dummy_ = Submodule_Stmt("submodule id) name")
     assert "Submodule_Stmt: 'submodule id) name'" in str(excinfo.value)
 
 
-def test_simple_error8(monkeypatch):
+def test_simple_error8():
     """Test the parsing of a submodule statement when there is a single
-    left hand bracket. The error generated here is unreachable if
-    splitparen works correctly so we need to monkeypatch.
+    left hand bracket.
 
     """
-
-    monkeypatch.setattr(
-        "fparser.common.splitline.splitparen", lambda x: ["", "(id", "name"]
-    )
     with pytest.raises(NoMatchError) as excinfo:
         dummy_ = Submodule_Stmt("submodule (id name")
     assert "Submodule_Stmt: 'submodule (id name'" in str(excinfo.value)
@@ -122,14 +112,12 @@ def test_simple_error8(monkeypatch):
 
 def test_splitparen_error(monkeypatch):
     """Test that if the first argument of the splitparen is not empty then
-    an error is returned.
+    an error is returned. Monkeypatch to force this error.
 
     """
-    # We must monkeypatch the splitparen function that has already been
-    # imported into the F2008 module.
-    monkeypatch.setattr(
-        "fparser.two.Fortran2008.Fortran2008.splitparen", lambda x: ["XXX", "", ""]
-    )
+    from fparser.common import splitline
+
+    monkeypatch.setattr(splitline, "splitparen", lambda x: ["XXX", "", ""])
     with pytest.raises(NoMatchError) as excinfo:
         dummy_ = Submodule_Stmt("submodule (id) name")
     assert "Submodule_Stmt: 'submodule (id) name'" in str(excinfo.value)
